Substitute some long/int with fixed-size data types. Fixes a hang in pathfinder in 64-bit build, among other things.

+++ src/Libraries/2D/Source/vtab.h
@@ -38,6 +38,6 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #ifndef __VTAB_H
 #define __VTAB_H
 
-extern long *gr_make_vtab (grs_bitmap *bm);
+extern int32_t *gr_make_vtab (grs_bitmap *bm);
 
 #endif
